+++ test/unit/fuzz-tools.test.js
@@ -473,6 +473,7 @@ import * as intermacsProfileV494 from '../../lib/intermacs-profile-v494.js';
 import * as ranawatMyelopathyV495 from '../../lib/ranawat-myelopathy-v495.js';
 import * as lodwickGradeV496 from '../../lib/lodwick-grade-v496.js';
 import * as schobingerAvmV497 from '../../lib/schobinger-avm-v497.js';
+import * as narakasObppV498 from '../../lib/narakas-obpp-v498.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -891,6 +892,7 @@ const MODULES = {
   'ranawat-myelopathy-v495.js': ranawatMyelopathyV495,
   'lodwick-grade-v496.js': lodwickGradeV496,
   'schobinger-avm-v497.js': schobingerAvmV497,
+  'narakas-obpp-v498.js': narakasObppV498,
   'tb-testing.js': tbTesting,
 };